VC7.1: fixing some build errors and warnings
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@781 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/util/impl/thread_posix.inl
@@ -58,7 +58,7 @@ namespace util
 namespace impl
 {
 
-size_t numberOfProcessors()
+unsigned numberOfProcessors()
 {
 	// we'll need to cache this if we want this to ever work ...
 	
@@ -66,7 +66,7 @@ size_t numberOfProcessors()
 	CPU_ZERO(&mask);	
 	LASS_ENFORCE_CLIB(sched_getaffinity(0, sizeof(cpu_set_t), &mask));
 	
-	size_t count = 0;
+	unsigned count = 0;
 	int i = 0;
 	while (CPU_ISSET(i++, &mask))
 	{
@@ -275,7 +275,7 @@ private:
 /** @internal
  *  @ingroup Threading
  */
-void bindThread(pid_t pid, size_t processor)
+void bindThread(pid_t pid, unsigned processor)
 {
 	cpu_set_t mask;
 	CPU_ZERO(&mask);
